jmxtrans-core/src/main/java/com/googlecode/jmxtrans/model/output/support/MaxBatchSizeWriter.java
/**
 * The MIT License
 * Copyright © 2010 JmxTrans team
 *
 * Permission is hereby granted, free of charge, to any person obtaining a copy
 * of this software and associated documentation files (the "Software"), to deal
 * in the Software without restriction, including without limitation the rights
 * to use, copy, modify, merge, publish, distribute, sublicense, and/or sell
 * copies of the Software, and to permit persons to whom the Software is
 * furnished to do so, subject to the following conditions:
 *
 * The above copyright notice and this permission notice shall be included in
 * all copies or substantial portions of the Software.
 *
 * THE SOFTWARE IS PROVIDED "AS IS", WITHOUT WARRANTY OF ANY KIND, EXPRESS OR
 * IMPLIED, INCLUDING BUT NOT LIMITED TO THE WARRANTIES OF MERCHANTABILITY,
 * FITNESS FOR A PARTICULAR PURPOSE AND NONINFRINGEMENT. IN NO EVENT SHALL THE
 * AUTHORS OR COPYRIGHT HOLDERS BE LIABLE FOR ANY CLAIM, DAMAGES OR OTHER
 * LIABILITY, WHETHER IN AN ACTION OF CONTRACT, TORT OR OTHERWISE, ARISING FROM,
 * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
 * THE SOFTWARE.
 */
package com.googlecode.jmxtrans.model.output.support;

import static com.google.common.collect.Iterables.partition;

import java.util.List;

import javax.annotation.Nonnull;

import com.googlecode.jmxtrans.model.OutputWriter;
import com.googlecode.jmxtrans.model.OutputWriterAdapter;
import com.googlecode.jmxtrans.model.Query;
import com.googlecode.jmxtrans.model.Result;
import com.googlecode.jmxtrans.model.Server;

public class MaxBatchSizeWriter<T extends OutputWriter> extends OutputWriterAdapter {

	private final int maxBatchSize;
	@Nonnull private final T target;

	public MaxBatchSizeWriter(int maxBatchSize, @Nonnull T target) {
		this.maxBatchSize = maxBatchSize;
		this.target = target;
	}

	@Override
	public void doWrite(@Nonnull Server server, @Nonnull Query query, @Nonnull Iterable<Result> results) throws Exception {
		for (List<Result> r : partition(results, maxBatchSize)) {
			target.doWrite(server, query, r);
		}
	}
}
